Move profile form to its own component. Add explicity type prop to button component
