+++ app/package.json
@@ -82,6 +82,7 @@
     "@types/webpack": "^4.4.9",
     "@types/webpack-merge": "^4.1.3",
     "@types/webpack-node-externals": "^1.6.3",
+    "@types/yup": "^0.26.23",
     "assets-webpack-plugin": "^3.9.5",
     "babel-loader": "^8.0.0-beta.4",
     "babel-plugin-lodash": "^3.3.4",
